+++ modules/agent-engine-nightly/main.tf
@@ -15,7 +15,8 @@
  */
 
 locals {
-  identity_type = lookup(var.spec, "identity_type", "SERVICE_ACCOUNT")
+  is_container_spec_deployment = var.spec != null && lookup(var.spec, "container_spec", null) != null
+  identity_type                = lookup(var.spec, "identity_type", "SERVICE_ACCOUNT")
 
   identity_prefixes = {
     "SERVICE_ACCOUNT" = "serviceAccount:"
@@ -23,6 +24,58 @@ locals {
   }
 
   member_prefix = lookup(local.identity_prefixes, local.identity_type, "")
+}
+
+resource "random_id" "suffix" {
+  count       = local.is_container_spec_deployment ? 1 : 0
+  provider    = random
+  byte_length = 4
+}
+
+resource "google_vertex_ai_reasoning_engine" "tenant_mds" {
+  count        = local.is_container_spec_deployment ? 1 : 0
+  provider     = google-nightly
+  display_name = "${var.display_name}-mds-${random_id.suffix[0].hex}"
+  description  = "Metadata agent to get tenant project service account"
+  region       = var.region
+  project      = var.project_id
+
+  spec {
+    source_code_spec {
+      inline_source {
+        source_archive = "H4sIAAAAAAAAA+2VXW+bMBSGuV1+hcVVKg0DSQAJKdK6Le0qdWmlaRfVNEUenBAawNQ2UdKq/30mhKXdh7KL0WrqeW6MP8+xX16b2kbnOJrA8+rSDTznYdliuN7AGfqO4/uu4bgD33UM4nWfmmFUUjFBiBGzVG7K5YJfs9+OO9T/n0Lt0+nnnC1hnmbQUYxaYH80+qP+A3envxcEw2Gg9R/5gdbf6SifR7x0/S8/XEyvQsKyjJQsWrIEej1dCfe13UfYewXFiry7uLw6OTufzN6ffTp+ez4Zu6Q+QCu6Xc0JpXYey5keXaiZFBHVXTS5JTkoFjPFmh5aboiAmyoVkOuqpGqtnvscXirUbiXpLsYh//tesPe/N6j9P3QG6P+nIMqYlOTjzp/H9b8Q9nqExDAnNxWITV9CNj8KdRMhaV5yobbeBanktq0SGRkTc6FUGdp2a3SacJ5kQNNCgShYZkc8LysFbSB75dql4NcQKbuochBpZO3qVhqb25WV2DRhawTIkhcSdKw2PE1A9XX412QBLAYhx3dmu751krEVF2ZIzNNtJub90S9rUaE1hdmci5n+CVQl+w/HqEoUZG5KEKs0AuvuxywFa3X/pm4HkYGUluDfuKrTj2nKcprsprAo4pW2lt56syFYR1AqMtkWKS/Cn6NNeaHvW8G5ai5KvdlHyvSP/rn+1O7+Jj7kf2fk7/3vD7X/fdfF9/9JaHxqRRmvYoulZcaU9kP+pXnCoUjSAuTX584SQRAEQRAEQRAEQRAEQRAEQRAEQZC/4TsjV0RZACgAAA=="
+      }
+
+      python_spec {
+        entrypoint_module = "agent"
+        entrypoint_object = "root_agent"
+      }
+    }
+  }
+}
+
+data "google_vertex_ai_reasoning_engine_query" "tenant_mds" {
+  count               = local.is_container_spec_deployment ? 1 : 0
+  provider            = google-nightly
+  project             = var.project_id
+  region              = var.region
+  reasoning_engine_id = google_vertex_ai_reasoning_engine.tenant_mds[0].name
+  depends_on          = [google_vertex_ai_reasoning_engine.tenant_mds]
+}
+
+resource "google_project_iam_member" "vertex_ar_reader" {
+  count      = local.is_container_spec_deployment ? 1 : 0
+  project    = data.google_project.project.project_id
+  role       = "roles/artifactregistry.reader"
+  member     = "serviceAccount:service-${data.google_project.project.number}@gcp-sa-aiplatform-re.iam.gserviceaccount.com"
+  depends_on = [data.google_vertex_ai_reasoning_engine_query.tenant_mds]
+}
+
+resource "google_project_iam_member" "tenant_ar_reader" {
+  count   = local.is_container_spec_deployment ? 1 : 0
+  project = data.google_project.project.project_id
+  role    = "roles/artifactregistry.reader"
+  member  = "serviceAccount:${jsondecode(data.google_vertex_ai_reasoning_engine_query.tenant_mds[0].output).output}"
 }
 
 resource "google_vertex_ai_reasoning_engine" "main" {
@@ -34,7 +87,9 @@ resource "google_vertex_ai_reasoning_engine" "main" {
 
   depends_on = [
     google_project_iam_member.aiplatform_roles,
-    var.module_depends_on
+    var.module_depends_on,
+    google_project_iam_member.vertex_ar_reader,
+    google_project_iam_member.tenant_ar_reader
   ]
 
   dynamic "encryption_spec" {
@@ -59,6 +114,13 @@ resource "google_vertex_ai_reasoning_engine" "main" {
           pickle_object_gcs_uri    = lookup(package_spec.value, "pickle_object_gcs_uri", null)
           python_version           = lookup(package_spec.value, "python_version", null)
           requirements_gcs_uri     = lookup(package_spec.value, "requirements_gcs_uri", null)
+        }
+      }
+
+      dynamic "container_spec" {
+        for_each = lookup(spec.value, "container_spec", null) == null ? [] : [spec.value.container_spec]
+        content {
+          image_uri = lookup(container_spec.value, "image_uri", null)
         }
       }
 
@@ -169,12 +231,6 @@ resource "google_vertex_ai_reasoning_engine" "main" {
       }
     }
   }
-}
-
-resource "time_sleep" "wait_for_auto_registration_for_agent_in_registry" {
-  # Trigger the sleep only after the reasoning engine is fully created
-  depends_on      = [google_vertex_ai_reasoning_engine.main]
-  create_duration = "10s"
 }
 
 resource "google_project_service_identity" "aiplatform_identity" {