+++ modules/agent-engine-nightly/versions.tf
@@ -28,9 +28,9 @@ terraform {
       source  = "hashicorp/google-nightly"
       version = "2026.4.8-7.27.0"
     }
-    time = {
-      source  = "hashicorp/time"
-      version = ">=0.13.1"
+    random = {
+      source  = "hashicorp/random"
+      version = ">=3.9.0"
     }
   }
   required_version = ">= 1.3"