+++ modules/model-armor-template/versions.tf
@@ -23,6 +23,6 @@ terraform {
     }
   }
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-vertex-ai:model-armor-template/v6.0.0"
+    module_name = "blueprints/terraform/terraform-google-vertex-ai:model-armor-template/v7.1.0"
   }
 }